% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_index.R
\name{fraction_index}
\alias{fraction_index}
\title{Compute the fraction of time spent in the right side for each period}
\usage{
fraction_index(pos = NULL, min_exits_required = NA, min_length = 10)
}
\arguments{
\item{pos}{A character vector encoding the position of the fly in the chamber (L/D/R)}

\item{min_required}{Number of exits required to compute the index}
}
\value{
Numeric ranging 0-1 0 shows total aversion for odour on right side, and 1 total preference. -1 is return if min_required is not reached
}
\description{
Compute the fraction of time spent in the right side for each period
}
\examples{
pos <- c('L','L','L','D','D','D','L','L',"L",'D','D','R','R','R','D')
preference_index(pos)
}
