% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Cardiovascular disease impact on the mortality of COVID-19}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
data("mortality")
}
\description{
This is a dataset containing several effect estimates and their standard errors
for the impact of cardivascular disease on the mortality of COVID-19 in the literature.
}
\examples{
data("mortality")
}
\keyword{datasets}
