% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_experiment.R
\name{read_experiment}
\alias{read_experiment}
\title{Read experiment data (odors and lemdt_results csv)}
\usage{
read_experiment(self)
}
\description{
Read experiment data (odors and lemdt_results csv)
}
