% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp_filtered.R
\docType{data}
\name{RPP_filtered}
\alias{RPP_filtered}
\title{Filtered RPP data}
\format{
An object of class \code{data.frame} with 73 rows and 5 columns.
}
\usage{
data("RPP_filtered")
}
\description{
Filtered RPP data
}
\examples{
data("RPP_filtered")
}
\keyword{datasets}
