% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_prp.R
\name{prior_prp}
\alias{prior_prp}
\title{Prior Predictive Replication p-value Calculation}
\usage{
prior_prp(
  beta,
  se,
  r_vec = c(0, 8e-04, 0.006, 0.024),
  test = "two_sided",
  report_PI = F
)
}
\arguments{
\item{beta}{A vector, containing the estimates in the original study and the
replication study.}

\item{se}{A vector, containing the standard errors of the estimates in the original
study and thereplication study.}

\item{r_vec}{A vector, defining the prior reproducible model. Each r value
corresponds to a probability.}

\item{test}{A string, determining which test statistics to use. If not sepecified,
the default two-sided one will be used.}

\item{report_PI}{A boolean, denoting whether the 95\% predictive interval for
the estimates be reported or not. This is option is only valid for two-sided
test statistics.}
}
\value{
A list with the following components:
\item{grid}{ The grid points for the hyperparameters.}
\item{test_statistics}{ The test statistics used in calculating the replication p-value.}
\item{pvalue}{ The resulting prior predictive replicaiton p-value.}
\item{predictive_interval}{The 95\% predictive interval if required.}
}
\description{
Assessing the prior predictive distribution and calculating the replication
p-value based on it.
}
